reduce unused code
