+++ index.html
@@ -5,6 +5,8 @@
   <meta charset="UTF-8" />
   <link rel="icon" href="/icon.ico" type="image/x-icon"/>
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <meta name="keywords" content="todo" >
+  <meta name="description" content="todo app you can add your todos" >
   
   <link rel="preconnect" href="https://fonts.googleapis.com">
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -13,7 +15,7 @@
     rel="stylesheet">
   <link href="https://fonts.googleapis.com/css2?family=Great+Vibes&display=swap" rel="stylesheet">
   
-  <title>Todos </title>
+  <title> Todos </title>
 </head>
 
 <body>
