+++ src/router/router-context.jsx
@@ -4,24 +4,21 @@ import { createBrowserRouter, createRoutesFromElements, Route, } from 'react-rou
 import RootLayout from '../layouts/rootLayout';
 import CircularProgress from '@mui/material/CircularProgress';
 
+// Lazy Load To improeve perormance
 const Login = React.lazy(() => import("../pages/login"));
 const Home = React.lazy(() => import("../pages/home"));
 const Register = React.lazy(() => import("../pages/register"));
 const UserHome = React.lazy(() => import("../pages/user-home"));
 
-// import Home from '../pages/home';
-// import Login from '../pages/login';
-// import Register from '../pages/register';
-// import UserHome from '../pages/user-home';
 import NotFound from '../pages/not-found';
 import PrivateRoute from './private.route';
 
 
 
-// Defining our custom router
+
 const RouterContext = createBrowserRouter(
   createRoutesFromElements(
-    // delcrere your routes
+
 
     <Route path='/' element={<RootLayout />}>
 
@@ -31,7 +28,7 @@ const RouterContext = createBrowserRouter(
 
       <Route path='register' element={<Suspense fallback={<CircularProgress />}><Register /></Suspense>} />
 
-      <Route path='user-home' element={<Suspense fallback={<CircularProgress />}><PrivateRoute><UserHome /></PrivateRoute></Suspense>} />
+      <Route path='todo-list-user' element={<Suspense fallback={<CircularProgress />}><PrivateRoute><UserHome /></PrivateRoute></Suspense>} />
 
       <Route path='*' element={<NotFound />} />
     </Route>
